// src/types/searchTypes.ts
export interface SearchTypes {
  isLoading: boolean; // ローディング管理用
  error: string;
  exactMode: boolean; // デフォルトは部分一致検索でtoggleさせる
  // eslint-disable-next-line camelcase
  number_1: string;
  // eslint-disable-next-line camelcase
  number_2: string;
  // eslint-disable-next-line camelcase
  number_3: string;
  // eslint-disable-next-line camelcase
  number_4: string;
  // eslint-disable-next-line camelcase
  number_5: string;
  // eslint-disable-next-line camelcase
  number_6: string;
  // eslint-disable-next-line camelcase
  number_7: string;
  // eslint-disable-next-line camelcase
  bonus_number1: string;
  // eslint-disable-next-line camelcase
  bonus_number2: string;
  // eslint-disable-next-line camelcase
  lottery_number: string;
}